made path intersection calculation work
