+++ lib/intersect.js
@@ -147,7 +147,6 @@ module.exports = function() {
 
         if (ip1 !== null && ip2 !== null) {
             var method;
-
             if (method = composedShapeMethods[ip1.type]) {
                 result = method(ip1, ip2, m1, m2);
             }
@@ -312,7 +311,6 @@ module.exports = function() {
             var pathParams = path.params[0];
             var inter0;
             var previnter;
-
             for (var inter, i = 0; i < pathParams.length; i++) {
                 inter = intersect(pathParams[i], shape, m1, m2);
                 if (!inter0)
